moved all logic from controller to service
