+++ appinfo/routes.php
@@ -27,6 +27,6 @@ return [
 		['name' => 'page#index', 'url' => '/{group}', 'verb' => 'GET', 'postfix' => 'group'],
 		['name' => 'page#index', 'url' => '/{group}/{contact}', 'verb' => 'GET', 'postfix' => 'group.contact'],
 		['name' => 'social_api#update_contact', 'url' => '/api/v1/social/avatar/{network}/{addressbookId}/{contactId}', 'verb' => 'GET'],
-		['name' => 'social_api#set_config', 	'url' => '/api/v1/social/config/{key}', 'verb' => 'POST'],
+		['name' => 'social_api#set_app_config', 'url' => '/api/v1/social/config/{key}', 'verb' => 'POST'],
 	]
 ];
