+++ lib/Controller/PageController.php
@@ -30,7 +30,7 @@ use OCP\IConfig;
 use OCP\IRequest;
 use OCP\L10N\IFactory;
 use OCP\Util;
-use OCA\Contacts\Controller\SocialApiController;
+use OCA\Contacts\Service\SocialApiService;
 
 class PageController extends Controller {
 
@@ -45,23 +45,22 @@ class PageController extends Controller {
 	/** @var IFactory */
 	private $languageFactory;
 
-	/** @var SocialApiController */
-	private $socialApi;
+	/** @var SocialApiService */
+	private  $socialApiService;
 
 	public function __construct(string $appName,
-								IRequest $request,
-								IConfig $config,
-								IInitialStateService $initialStateService,
-								IFactory $languageFactory,
-								SocialApiController $socialApi) {
+					IRequest $request,
+					IConfig $config,
+					IInitialStateService $initialStateService,
+					IFactory $languageFactory,
+					SocialApiService $socialApiService) {
 		parent::__construct($appName, $request);
 
 		$this->appName = $appName;
 		$this->config = $config;
 		$this->initialStateService = $initialStateService;
 		$this->languageFactory = $languageFactory;
-
-		$this->socialApi = $socialApi;
+		$this->socialApiService = $socialApiService;
 	}
 
 	/**
@@ -73,7 +72,7 @@ class PageController extends Controller {
 	public function index(): TemplateResponse {
 		$locales = $this->languageFactory->findAvailableLocales();
 		$defaultProfile = $this->config->getAppValue($this->appName, 'defaultProfile', 'HOME');
-		$supportedNetworks = $this->socialApi->getSupportedNetworks();
+		$supportedNetworks = $this->socialApiService->getSupportedNetworks();
 
 		$this->initialStateService->provideInitialState($this->appName, 'locales', $locales);
 		$this->initialStateService->provideInitialState($this->appName, 'defaultProfile', $defaultProfile);
