+++ tests/unit/Controller/PageControllerTest.php
@@ -29,7 +29,7 @@ use PHPUnit\Framework\MockObject\MockObject;
 use OCP\IInitialStateService;
 use OCP\IRequest;
 use OCP\L10N\IFactory;
-use OCA\Contacts\Controller\SocialApiController;
+use OCA\Contacts\Service\SocialApiService;
 use ChristophWurst\Nextcloud\Testing\TestCase;
 
 class PageControllerTest extends TestCase {
@@ -39,23 +39,26 @@ class PageControllerTest extends TestCase {
 	/** @var IRequest|MockObject */
 	private $request;
 
+	/** @var IConfig|MockObject*/
+	private  $config;
+
 	/** @var IInitialStateService|MockObject */
 	private $initialStateService;
 
 	/** @var IFactory|MockObject */
 	private $languageFactory;
 
-	/** @var IConfig|MockObject*/
-	private  $config;
+	/** @var SocialApiService|MockObject*/
+	private  $socialApi;
 
 	public function setUp() {
 		parent::setUp();
 
 		$this->request = $this->createMock(IRequest::class);
+		$this->config = $this->createMock(IConfig::class);
 		$this->initialStateService = $this->createMock(IInitialStateService::class);
 		$this->languageFactory = $this->createMock(IFactory::class);
-		$this->config = $this->createMock(IConfig::class);
-		$this->socialApi = $this->createMock(SocialApiController::class);
+		$this->socialApi = $this->createMock(SocialApiService::class);
 
 		$this->controller = new PageController(
 			'contacts',
